requirements.txt
pyspark==3.5.1
psycopg2-binary==2.9.9
vaderSentiment==3.3.2
requests==2.31.0
pyyaml==6.0.1
